[rename & correct] 파일이름 오타 수정
